DownwardAPI: Write netdev Mac address
Users of the network deviceinfo downwardAPI might need mac address
information as well.

That is the case of vDPA devices, which are assigned a mac address by
the sriov-network-operator right when they are provisioned.

Network binding plugins will need to fill that information in the
domainXML, if we want to avoid assigning a random mac address to those
devices (again), which could break network traffic.

This commit extends the downwardapi's networkinfo structure, and fills
it with the information obtained from multus.

Note that this could be a breaking change for those network binding
plugins which are expecting to unmarshall an old networkinfo Interfaces
struct. In other words, this changes the API between the downwardAPI and
network binding plugins.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>