vdpa: Configure a MAC address by default
Usually MAC addresses are assigned to vdpa devices right when they are
provisioned. That information is present in the multus network info.
Instead of relying on libvirt, which will generate a random one in the
absence of an explicitly configured MAC address, this will set the
original MAC address of the VDPA by default, except for when the user
set another one explicitly.

This helps fixing the issue, in which libvirt would assign a random mac
address to the network device, and break its traffic.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>