Network binding plugin: Add MemoryLockRequirements field
Some network binding plugins need to make sure that the memory lock
rlimits are equal or greater than the ones libvirt expects.

Just as sriov, some network interfaces require to lock the entirety of
the guest memory, and an extra bit sometimes to function properly. There
are other options that let cluster-admins configure the expected
resource overhead of network binding plugins, including memory. But that
does not impact the virt-launcher memory lock rlimits and it does affect
the amount of VMs that can run on a node, as that overhead limits the
amount of schedulable pods.

In other words, there is not a way of configuring process memory lock
rlimits according to network binding plugin needs currently.

This commit introduces the MemoryLockRequirements so cluster-admins can
set network binding plugin memory lock rlimits expectations. This is
configured through two parameters: lockGuestMemory and offset.
LockGuestMemory is a boolean flag that defines if the guest memory is
taken into account in the memory lock rlimits per each interface
attached to the vmi. Offset is a resource.Quantity that will add a
general overhead to the rlimit for all the interfaces that relate to a
certain network binding plugin type.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>