virt-handler: Refactor AdjustResources and AdjustQemuProcessMemoryLimits
Instead of returning early if none of the conditions are met, compute
memLockSize only if those conditions are met, and return if it is still
zero after that.

The behavior is not altered. This makes it easier for later to extend
memLockSize from other sources if needed, making the existing behavior
and any new possible mechanism independent from each other.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>